Bug:            #721 Submitted by:   wb@willbarton.com (Will Barton) Reviewed by: Approved by: Obtained from:
This patch makes a few corrections to the mirror site list documentation I submitted a few days
ago, based on suggestions made by Dr. Ernie.  (1) Be a bit more specific about the nature of the
various tokens in the example in Chapter 2, and (2) Be a bit more specific about
master_sites.mirror_subdir in the example in Chapter 2.


git-svn-id: http://svn.macports.org/repository/macports/trunk@3093 d073be05-634f-4543-b044-5fe20cf6d1d6