misc: Remove Tiger apple-gcc42 recipie
This was a snowball that really shouldn't be in Portfiles.

Tiger is old and crufty, and there should be a better approach for this.
The immediate issue that sparked these changes was a bug in an earlier
version of ld64 which caused a snowball.  That has since been fixed.

There may be lingering issues with some ports requiring a rebuild if
ld_classic doesn't see its symbols, but that should just need a
revbump where appropriate (in the port that provides the symbols).


git-svn-id: http://svn.macports.org/repository/macports/trunk@92250 d073be05-634f-4543-b044-5fe20cf6d1d6