darwintrace: fix expansion of multiple symlinks
Expanding a series of symlinks longer than a single symlink would previously
lead to re-use of a buffer that was used to store the result of readlink(2)
while the contents of the buffer were still in use in the pathComponent array.

This change refactors the path parsing and handling completely such that the
components of the normalized path will always be stored in the same buffer (and
other components will be copied there when needed), avoiding this problem.

Fixes the configure phase of stellarium and automoc (if you don't have
DARWINTRACE_DEBUG enabled).


git-svn-id: http://svn.macports.org/repository/macports/trunk@114096 d073be05-634f-4543-b044-5fe20cf6d1d6