Perl graveyard: Inter all retired p5-* subports (#46322)
Sorry for the cemetery jokes. (P.S., not really sorry.)

This adds p5.8, p5.10, p5.12, and p5.14 ports back to the port index
after r130129 and replaces them with the appropriate p5.16 port. The
following ports are omitted because they handle replacement themselves
in one way or another:
- p5-app-ack
- p5-catalyst-plugin-unicode-encoding
- p5-class-mop
- p5-csp
- p5-email-mime-modifier
- p5-file-slurp-unicode
- p5-graveyard
- p5-json-pp
- p5-test-tester
- p5-test-use-ok
- p5-vcp-autrijus-snapshot
- p5-vcp-dest-svk
- p5-vcp-source-cvsbk

Some of the restored subports might be "new", if they hadn't existed
prior to r130129. This is fine, since they can't be installed anyway.


git-svn-id: http://svn.macports.org/repository/macports/trunk@130773 d073be05-634f-4543-b044-5fe20cf6d1d6