cgal: remove demo variant; closes #44727
This variant hasn't worked correctly for a long time now. With newer versions
of cgal, it has been impossible to even build correctly. The design of the cgal
examples and demos now has a meaning of "install cgal first, then extract the
demos somewhere and run those separate." Therefore, we just remove the useless
variant.


git-svn-id: http://svn.macports.org/repository/macports/trunk@130957 d073be05-634f-4543-b044-5fe20cf6d1d6