base: Avoid overlap between existsuser/existsgroup error code and root/wheel's UID/GID, closes #45737
Return -1 from existsuser and existsgroup when a user or group does not exist.
Because these commands return the UID or GID in case of success, they could not
be used to check for the existence of the root user or the wheel group (UID/GID
0).

Since existsuser and existsgroup are used in adduser and addgroup in
port1.0/portutil.tcl, putting add_user root into a Portfile would make MacPorts
trash the system's root user by replacing it with a new user with a non-zero
UID, effectively stripping the root user from its privileges.

This caused a problem in the dbus port when installed in a root MacPorts
installation with the +no_root variant, which is explained in #45737.


git-svn-id: http://svn.macports.org/repository/macports/trunk@135851 d073be05-634f-4543-b044-5fe20cf6d1d6