certsync: Avoid segfault in absence of kSecTrustSettingsResult, closes #47906
Root certificates apparently sometimes do not have a kSecTrustSettingsResult,
and the absence should be treated as kSecTrustSettingsResultTrustRoot. This
change implements that.

Additionally, this silences a few warnings emitted by clang about functions
that are never NULL (at least not on the platform you're compiling for). Since
these checks are required for other platforms, employ the address-of operator
as suggested by clang to turn off the warning.


git-svn-id: http://svn.macports.org/repository/macports/trunk@136958 d073be05-634f-4543-b044-5fe20cf6d1d6