base: tracelib: Add process tracking & killing
Some build systems (*cough* xcodebuild *cough*) start long-running daemon
process that do terminate after the build is finished. Because these processes
use file descriptors, they are instrumented by darwintrace and end up in an
inconsistent state when the tracelib server terminates at the end of the build.
These processes then commonly hang around as quasi-zombies and cannot be killed
but using SIGKILL.

Avoid this problem by tracking which process belongs to a incoming socket in
tracelib and sending these process both SIGTERM and SIGKILL after the main loop
terminates.

NOTE: This change implements the data structures and operations required to do
the tracking, but does not actually fill and drain the list of open sockets.
This is because I did some refactoring while writing these changes and the
control flow path that adds and removes entries from the list can not be easily
separated (short of a complete re-implementation against an older version that
will be replaced anyway) from other refactoring.


git-svn-id: http://svn.macports.org/repository/macports/trunk@140639 d073be05-634f-4543-b044-5fe20cf6d1d6