base: tracelib: Refactor, enable process tracking
Refactor both the synchronization and error handling in tracelib to
clean up some of the copy-n-paste cruft that has accumulated over time.
Additionally, add the code that fills and depletes the list of connected
sockets and their processes at the remote ends to implement cleaning
them up after the build finishes.

This simplifies how the synchronization between "tracelib run" and
"tracelib closesocket" works. "tracelib run" now uses a simple event
loop idiom that receives and process asynchronous events and terminates
at the request of "tracelib closesocket" using a well-defined channel (a
selfpipe, since it's the only portable method to do this without kqueue
user events, which are not available on older OS X versions).


git-svn-id: http://svn.macports.org/repository/macports/trunk@140641 d073be05-634f-4543-b044-5fe20cf6d1d6