gmp: Link with -no-undefined
I encountered an odd error on a Leopard VM where some libgmp symbols were
missing in the library.  Had the library been built with -no-undefined, the
issue would have been caught at link time.  Oddly, rebuilding (with
-no-undefined) resulted in the symbols being present.  I highly doubt that
the flag is the fix for the problem.  It's likely that toolchain changes
in the interim fixed the issue for me, but I'm adding the flag to protect
users from potentially hitting fallout from the same issue in the future.


git-svn-id: http://svn.macports.org/repository/macports/trunk@142957 d073be05-634f-4543-b044-5fe20cf6d1d6