base: stop recording mtimes and names when compressing manpages
The manpage modification times get stored into a header of the gzip format when
compressing them. This leads to non-reproducible builds, because the file
modification times of those manpages differ between builds.

Of course the file modification times themselves are an issue that still needs
to be fixed because the generated tarballs will not be reproducible otherwise,
but this change itself is still worthwhile.


git-svn-id: http://svn.macports.org/repository/macports/trunk@143068 d073be05-634f-4543-b044-5fe20cf6d1d6