pymol-devel: Declare obsolete.
There is a pymol port, which is a newer version of the same software. It is not clear what, if any, purpose pymol-devel serves, though years ago it was a newer version than the pymol port (no clear reason was stated at the time for why this was useful, either). pymol-devel has had no maintainer since 3 years ago (r105836). Since that time it has not been updated, and only received maintenance commits. The confusing nature of this port was discussed in this ticket a year and a half ago: #42860


git-svn-id: http://svn.macports.org/repository/macports/trunk@144344 d073be05-634f-4543-b044-5fe20cf6d1d6