nginx: 1.9.11
Adds dynamic module support and TCP support in the DNS resolver. Documentation
for dynamic modules is at
  https://www.nginx.com/resources/wiki/extending/converting/
and we might want to convert some modules to dynamic ones in our packaging.
However, modules must change to support this, and at this point, few probably
have. Additionally, dynamic modules must be enabled in the configuration, which
is a change of behavior for our users unless we find a way to do it
automatically for them. For these reasons, we don't use dynamic module support
yet.


git-svn-id: http://svn.macports.org/repository/macports/trunk@145690 d073be05-634f-4543-b044-5fe20cf6d1d6