pextlib/system: ignore signals SIGINT/SIGQUIT
When a SIGINT/SIGQUIT was generated by the user during execution of a program,
both the program and the waiting parent tclsh process got interrupted by
a signal. However, the handler in the tclsh only generated an unpleasant error
message that a SIGINT/SIGQUIT was received.

With this change, only the child will handle the signal, hopefully ending its
execution. If the program handles SIGINT/SIGQUIT, it might still just terminate
with a normal exit code, so there is no chance for us to determine that it was
killed by a signal.


git-svn-id: http://svn.macports.org/repository/macports/trunk@147224 d073be05-634f-4543-b044-5fe20cf6d1d6