quassel: use qt5
Especially on high-DPI displays, Qt5 performs a lot better than Qt4.

Unfortunately, our list of dependencies supporting Qt5 is not very long at
the moment; notably we do not have phonon or dbusmenu-qt for Qt5 (although
the latter seems to be used for clickable menus which seems to work without
DBus on OS X). For QCA we only have a variant for Qt5 support, which would
require using the active_variants PortGroup and manual user interaction.
Since QCA is only needed for client-side encrypted IRC communication, which
isn't widely used, I think we can get away without it for now.


git-svn-id: http://svn.macports.org/repository/macports/trunk@149653 d073be05-634f-4543-b044-5fe20cf6d1d6