lang/sbcl: change linking options for XCode 8
<https://trac.macports.org/attachment/ticket/51733/>

This change is required as the Xcode 8 linker now defaults to PIE
binaries regardless of the deployment target so an explicit
'-Wl,-no_pie' is now required. The rbcl program is incompatible with
ASLR and the previous upstream hack of setting the deployment target
to less than 10.7 doesn't prevent the PIE executables from being
created against the Xcode 8 linker. Tested on 10.11 with Xcode 8 beta.


git-svn-id: http://svn.macports.org/repository/macports/trunk@151301 d073be05-634f-4543-b044-5fe20cf6d1d6