portmgr: Fix PortIndex2MySQL #!, drop Makefile
The shebang of PortIndex2MySQL.tcl was hardcoded to /opt/local/bin/port-tclsh.
Use /usr/bin/env port-tclsh to make the script portable and obsolete the
Makefile (which didn't do anything other than copying the file for a while
now). Create a symlink so that use cases that expect the script to be available
as "PortIndex2MySQL" don't break.


git-svn-id: http://svn.macports.org/repository/macports/trunk@151423 d073be05-634f-4543-b044-5fe20cf6d1d6