base: porttrace: Do not store duplicate violations
The violations list became slow for large builds, which could be noticed
especially when software uses Qt (which creates a large number of false
positive violations in trace mode at the moment due to the use of directory
symlinks).

Replace the current list that kept duplicates with a list that will deduplicate
and automatically sort in O(log n).

Add a unit test to ensure that the implementation works as expected.


git-svn-id: http://svn.macports.org/repository/macports/trunk@151777 d073be05-634f-4543-b044-5fe20cf6d1d6